✨ ユーザー判定機能

+++ pages/api/item/delete/[id].js
@@ -5,8 +5,13 @@ import { ItemModel } from "../../../../utils/schemaModels";
 const deleteItem = async (req, res) => {
   try {
     await connectDB();
-    await ItemModel.deleteOne({ _id: req.query.id });
-    return res.status(200).json({ message: "アイテム削除成功" });
+    const singleItem = await ItemModel.findById(req.query.id);
+    if (singleItem.email === req.query.email) {
+      await ItemModel.deleteOne({ _id: req.query.id });
+      return res.status(200).json({ message: "アイテム削除成功" });
+    } else {
+      throw new Error();
+    }
   } catch (err) {
     return res.status(400).json({ message: "アイテム削除失敗" });
   }
